Fix master CI: check if coreneuron is loadable for given test
- In order to run python tests with coreneuron, we need to either
  use special (where coreneuron is linked) or have shared library
  build so that coreneuron can be loaded dynamically
- Some builds on gitlab also tests build with static libraries. So
  the test added in #2733 fails as we don't check if static builds
- In this PR we use existing mechanism/function to check if coreneuron
  is usable and then run test with coreneuron
